# apps/desktop/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "eigenheim engine",
    "version": "0.1.0"
  },
  "paths": {
...
    "/goals": {
      "get": {
        "summary": "List Goals",
        "description": "OKR tree: Objectives with KRs, each KR carrying live value, target,\ncomparison, status, progress, gap, trace_ref, and task_count.",
        "operationId": "list_goals_goals_get",
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Period"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GoalsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/objectives": {
      "get": {
        "summary": "List Objectives Endpoint",
        "description": "List Objectives (lightweight, no KR computation).",
        "operationId": "list_objectives_endpoint_objectives_get",
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Period"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ObjectiveSimpleOut"
                  },
                  "title": "Response List Objectives Endpoint Objectives Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Objective",
        "description": "Create an Objective.",
        "operationId": "create_objective_objectives_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObjectiveIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectiveSimpleOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/objectives/{objective_id}": {
      "get": {
        "summary": "Get Objective",
        "description": "Get one Objective with its KRs (live-computed).",
        "operationId": "get_objective_objectives__objective_id__get",
        "parameters": [
          {
            "name": "objective_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Objective Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectiveWithKRsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Objective",
        "description": "Update an Objective's title and/or period.",
        "operationId": "update_objective_objectives__objective_id__patch",
        "parameters": [
          {
            "name": "objective_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Objective Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ObjectiveUpdateIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ObjectiveSimpleOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Objective",
        "description": "Delete an Objective and all its KRs.",
        "operationId": "delete_objective_objectives__objective_id__delete",
        "parameters": [
          {
            "name": "objective_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Objective Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/key-results": {
      "post": {
        "summary": "Create Key Result",
        "description": "Create a KR under an Objective.",
        "operationId": "create_key_result_key_results_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KeyResultIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KeyResultOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/key-results/{kr_id}": {
      "get": {
        "summary": "Get Key Result",
        "description": "Get one KR with live-computed value, status, progress, gap, trace_ref.",
        "operationId": "get_key_result_key_results__kr_id__get",
        "parameters": [
          {
            "name": "kr_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kr Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KeyResultOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update Key Result",
        "description": "Update a KR's fields. Partial update: only supplied fields change.",
        "operationId": "update_key_result_key_results__kr_id__patch",
        "parameters": [
          {
            "name": "kr_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kr Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KeyResultUpdateIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KeyResultOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Key Result",
        "description": "Delete a KR.",
        "operationId": "delete_key_result_key_results__kr_id__delete",
        "parameters": [
          {
            "name": "kr_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kr Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeletedOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BoundLogicOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "version": {
            "type": "string",
            "title": "Version"
          },
          "usage_count": {
            "type": "integer",
            "title": "Usage Count"
          },
          "current_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Value"
          },
          "source_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Name"
          },
          "trace_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Ref"
          },
          "computed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Computed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "version",
          "usage_count"
        ],
        "title": "BoundLogicOut"
      },
...
      "DeletedOut": {
        "properties": {
          "deleted": {
            "type": "string",
            "title": "Deleted"
          }
        },
        "type": "object",
        "required": [
          "deleted"
        ],
        "title": "DeletedOut"
      },
...
      "GoalsOut": {
        "properties": {
          "objectives": {
            "items": {
              "$ref": "#/components/schemas/ObjectiveOut"
            },
            "type": "array",
            "title": "Objectives"
          },
          "period": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period"
          }
        },
        "type": "object",
        "required": [
          "objectives"
        ],
        "title": "GoalsOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KeyResultIn": {
        "properties": {
          "objective_id": {
            "type": "string",
            "title": "Objective Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "logic_id": {
            "type": "string",
            "title": "Logic Id",
            "default": ""
          },
          "target": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          },
          "comparison": {
            "type": "string",
            "title": "Comparison",
            "default": "gte"
          },
          "period": {
            "type": "string",
            "title": "Period",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "objective_id",
          "name"
        ],
        "title": "KeyResultIn"
      },
      "KeyResultOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "objective_id": {
            "type": "string",
            "title": "Objective Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "logic_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logic Id"
          },
          "target": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          },
          "comparison": {
            "type": "string",
            "title": "Comparison"
          },
          "period": {
            "type": "string",
            "title": "Period"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "progress": {
            "type": "number",
            "title": "Progress"
          },
          "gap": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gap"
          },
          "live_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Live Value"
          },
          "live_raw": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Live Raw"
          },
          "trace_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Ref"
          },
          "task_count": {
            "type": "integer",
            "title": "Task Count"
          },
          "logic": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BoundLogicOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "spark": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Spark"
          }
        },
        "type": "object",
        "required": [
          "id",
          "objective_id",
          "name",
          "comparison",
          "period",
          "created_at",
          "status",
          "progress",
          "task_count",
          "spark"
        ],
        "title": "KeyResultOut"
      },
      "KeyResultUpdateIn": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "logic_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logic Id"
          },
          "target": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target"
          },
          "comparison": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comparison"
          },
          "period": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period"
          }
        },
        "type": "object",
        "title": "KeyResultUpdateIn"
      },
...
      "ObjectiveIn": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "period": {
            "type": "string",
            "title": "Period",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "ObjectiveIn"
      },
      "ObjectiveOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "period": {
            "type": "string",
            "title": "Period"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "krs": {
            "items": {
              "$ref": "#/components/schemas/KeyResultOut"
            },
            "type": "array",
            "title": "Krs"
          },
          "on_track_count": {
            "type": "integer",
            "title": "On Track Count"
          },
          "total_kr_count": {
            "type": "integer",
            "title": "Total Kr Count"
          },
          "aggregate_progress": {
            "type": "number",
            "title": "Aggregate Progress"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "period",
          "source",
          "krs",
          "on_track_count",
          "total_kr_count",
          "aggregate_progress"
        ],
        "title": "ObjectiveOut"
      },
      "ObjectiveSimpleOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "period": {
            "type": "string",
            "title": "Period"
          },
          "source": {
            "type": "string",
            "title": "Source"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "period",
          "source"
        ],
        "title": "ObjectiveSimpleOut",
        "description": "Lightweight objective returned by GET /objectives (no KR computation)."
      },
      "ObjectiveUpdateIn": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "period": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period"
          }
        },
        "type": "object",
        "title": "ObjectiveUpdateIn"
      },
      "ObjectiveWithKRsOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "period": {
            "type": "string",
            "title": "Period"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "krs": {
            "items": {
              "$ref": "#/components/schemas/KeyResultOut"
            },
            "type": "array",
            "title": "Krs"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "period",
          "source",
          "krs"
        ],
        "title": "ObjectiveWithKRsOut",
        "description": "Single-objective response from GET /objectives/{id}."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}